Remove old MediaItemSource concept
