+++ pillarbox-demo/src/main/java/ch/srgssr/pillarbox/demo/ui/showcases/layouts/StoryViewModel.kt
@@ -10,7 +10,6 @@ import androidx.media3.common.C
 import androidx.media3.common.Player
 import ch.srgssr.pillarbox.core.business.tracker.DefaultMediaItemTrackerRepository
 import ch.srgssr.pillarbox.demo.shared.data.Playlist
-import ch.srgssr.pillarbox.demo.shared.di.PlayerModule
 import ch.srgssr.pillarbox.player.PillarboxPlayer
 import kotlin.math.ceil
 
@@ -20,7 +19,6 @@ import kotlin.math.ceil
  * 3 Players that interleaved DemoItems
  */
 class StoryViewModel(application: Application) : AndroidViewModel(application) {
-    private val mediaItemSource = PlayerModule.provideMixedItemSource(application)
     private val itemTrackerProvider = DefaultMediaItemTrackerRepository()
 
     /**
@@ -29,17 +27,14 @@ class StoryViewModel(application: Application) : AndroidViewModel(application) {
     private val players = arrayOf(
         PillarboxPlayer(
             context = application,
-            // mediaItemSource = mediaItemSource,
             mediaItemTrackerProvider = itemTrackerProvider
         ),
         PillarboxPlayer(
             context = application,
-            // mediaItemSource = mediaItemSource,
             mediaItemTrackerProvider = itemTrackerProvider
         ),
         PillarboxPlayer(
             context = application,
-            // mediaItemSource = mediaItemSource,
             mediaItemTrackerProvider = itemTrackerProvider
         )
     )
