--- pillarbox-player/src/main/java/ch/srgssr/pillarbox/player/data/MediaItemSource.kt
@@ -1,20 +1,0 @@
-/*
- * Copyright (c) SRG SSR. All rights reserved.
- * License information is available from the LICENSE file.
- */
-package ch.srgssr.pillarbox.player.data
-
-import androidx.media3.common.MediaItem
-
-/**
- * Media item source load MediaItem with a suspend function.
- */
-interface MediaItemSource {
-    /**
-     * Load media item from [mediaItem] in a suspend function
-     *
-     * @param mediaItem
-     * @return MediaItem buildUpon [mediaItem]
-     */
-    suspend fun loadMediaItem(mediaItem: MediaItem): MediaItem
-}
